refactor: remove `_app`, `_document`. and move it to `layout`
